Solution for problem 3 using trial division.

+++ include/problem_0003.h
@@ -1,6 +1,6 @@
 #ifndef PROBLEM_0003_H
 #define PROBLEM_0003_H
 
-int problem_0003(int composite);
+int problem_0003(long composite);
 
 #endif
